app/requirements.txt
anthropic>=0.40.0
streamlit>=1.29.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
